All functions working; pre-adding google maps

+++ client/src/components/Cards/ComplaintCard.js
@@ -2,40 +2,42 @@
 import { useContext, useState, useEffect } from "react";
 import { ComplaintTypesContext } from "../ComplaintTypesInfo";
 import { BuildingsContext } from "../BuildingsInfo";
+import { TenantComplaintContext } from "../TenantComplaintsContext";
 import { UserContext } from "../UserContext"
 import EditComplaint from "../EditComplaint";
 import Errors from "../Errors"
 
 
-function ComplaintCard({tenantComplaint, complaints, setComplaints}){
- 
-  
+function ComplaintCard({tenantComplaint}){
+  // console.log(tenantComplaint)
   const createdAt = new Date(tenantComplaint.created_at)
   const updatedAt = new Date(tenantComplaint.updated_at)
   const now = new Date()
   const timeLapsed = Math.floor(Math.abs(now - createdAt)/(1000 * 60 * 60 * 24))
   
   const[editWindowOn, setEditWindowOn] = useState(false)
-  const[complaint, setComplaint] = useState({tenantComplaint})
+  const[complaint, setComplaint] = useState(tenantComplaint)
   const[errors, setErrors] = useState([])
   const[isSuper, setIsSuper] = useState(false)
   const[compStatus, setCompStatus] = useState()
   
   const complaintTypes = useContext(ComplaintTypesContext)
-  
+  const [complaints, setComplaints] = useContext(TenantComplaintContext)
   const buildings = useContext(BuildingsContext)
   const [user] = useContext(UserContext)
-  console.log(complaints)
+
+  
   useEffect(()=>{
     setComplaint(tenantComplaint);
+    // setTypeId(tenantComplaint.complaint_id)
     if(user && user.buildings){
       setIsSuper(true)
     }
   },[user])
 
-  // useEffect(()=>{
-  //   setCompStatus(tenantComplaint.resolved)
-  // })
+  useEffect(()=>{
+    setCompStatus(tenantComplaint.resolved)
+  },[])
   
   let building
   
@@ -44,8 +46,11 @@ function ComplaintCard({tenantComplaint, complaints, setComplaints}){
   }
 
   let complaintType
+  
   if(complaint){
-    complaintType = complaintTypes.find(c=>c.id===complaint.id)
+ 
+    complaintType = complaintTypes.find(c=>c.id===tenantComplaint.complaint_id)
+
   }
 
   
@@ -60,37 +65,40 @@ function ComplaintCard({tenantComplaint, complaints, setComplaints}){
     
 
   function handleDelete(){
-    if(isSuper){
-      setErrors(["Only a tenant can delete a complaint."])
-    } else {
-      const deleteConfirm = window.confirm("Are you sure you want to delete this complaint? You CANNOT undo this action. 'OK' to confirm, or CANCEL.")
-      if(deleteConfirm){
-        fetch(`/tenant_complaints/${tenantComplaint.id}`,{
-        method: "DELETE"
-        })
-        .then((r)=>{
-          if(r.ok){
-           
-            const filteredComplaints = complaints.filter((comp)=>comp.id!==(parseInt(complaint.id)));
-            setComplaints(filteredComplaints)
-          } else {
-              r.json().then((err)=> setErrors(err.errors))
-            }
-            
+    if(!user){
+      setErrors(["You must be logged in to edit/update/delete complaints."]);
+    } else{
+      if(isSuper){
+        setErrors(["Only a tenant can delete a complaint."])
+      } else {
+        const deleteConfirm = window.confirm("Are you sure you want to delete this complaint? You CANNOT undo this action. 'OK' to confirm, or CANCEL.")
+        if(deleteConfirm){
+          fetch(`/tenant_complaints/${tenantComplaint.id}`,{
+          method: "DELETE"
+          })
+          .then((r)=>{
+            if(r.ok){
+             
+              const filteredComplaints = complaints.filter((comp)=>comp.id!==(parseInt(complaint.id)));
+              setComplaints(filteredComplaints)
+            } else {
+                r.json().then((err)=> setErrors(err.errors))
+              }
+              
           })
         }
       }
-    
-    }  
+    }
+  }  
  
-  
+
   if(tenantComplaint!==undefined){
     return(
       <div id="complaint-card">
-        {complaintType ? <h3>{complaintType.complaint_type}</h3> : ""}
+        <h3>{complaint.complaint_type}</h3>
         {building ? <h4>{building.address}</h4>: "" }
-        <div>Unit #: {tenantComplaint.unit}</div>
-        <div>Tenant notes: {tenantComplaint.tenant_notes}</div>
+        <div>Unit #: {complaint.unit}</div>
+        <div>Tenant notes: {complaint.tenant_notes}</div>
         <div>Super notes: {complaint.super_notes}</div>
         <div>Opened: {createdAt.toString()}</div>
         <div>Updated: {updatedAt.toString()}</div>
@@ -101,7 +109,7 @@ function ComplaintCard({tenantComplaint, complaints, setComplaints}){
           <button onClick={handleDelete}>Delete Complaint</button>
         </div>
         <div id="edit-window">
-          {editWindowOn ? <EditComplaint complaintId={tenantComplaint.id} setComplaint={setComplaint} setComplaints={setComplaints} isSuper={isSuper} setIsSuper={setIsSuper}/> : ""}
+          {editWindowOn ? <EditComplaint complaintId={tenantComplaint.id} complaint={complaint} setComplaint={setComplaint} setComplaints={setComplaints} isSuper={isSuper} setIsSuper={setIsSuper}/> : ""}
         </div>
         <div>
           <Errors errors={errors}/>
